% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{The bootstrap function creates a confidence interval based on a given data set.}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of iterations}

\item{x}{A dataframe or data set}

\item{fun}{Which type of function you would like to apply to the data, ex. "mean"}

\item{alpha}{A number, such that (1 - alpha) is the percentage confidence}

\item{cx}{A number, for applying text size}

\item{...}{Extra graphical parameters passed on to the hist() function}
}
\value{
Returns a histogram of the distribution of the sampled data from the population. On the histogram is a labeled confidence interval.
}
\description{
The bootstrap function creates a confidence interval based on a given data set.
}
\examples{
\dontrun{myboot2(iter = 10000, x = ddt$DDT, fun = "var", alpha = 0.2)}
}
