% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{This function creates a random sample from a uniform distribution and creates plots based on the shape of the sample distribution.}
\usage{
myclt(n = 1, iter = 10000, a = 0, b = 5)
}
\arguments{
\item{n}{Number of trials}

\item{iter}{Number of iterations}

\item{a}{A number, representing the minimum value of the sample distribution.}

\item{b}{A number, representing the maximum value of the sample distribution.}
}
\value{
Returns a plot of the shape of the sample distribution taken randomly from a uniform distribution.
}
\description{
This function creates a random sample from a uniform distribution and creates plots based on the shape of the sample distribution.
}
\examples{

\dontrun{myclt(n = 20, iter = 100000, a = 0, b = 10)}
}
