% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myddt.R
\name{myddt}
\alias{myddt}
\title{This function plots length vs. width of a given species.}
\usage{
myddt(df, SPECIES)
}
\arguments{
\item{df}{A data frame, must be the DDT data frame.}

\item{SPECIES}{A species listed in the DDT data frame.}
}
\value{
Returns a plot of the length vs. the weight of a given species, color coded by river. Also returns the entire data frame, the subsetted data frame, and a RIVER relative frequency table.
}
\description{
This function plots length vs. width of a given species.
}
\examples{
\dontrun{myddt(df = DDT.csv, SPECIES = "CCATFISH")}

}
