% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{mymaxlikg Creates a graphical likelihood for a probability.}
\usage{
mymaxlikg(lfun, theta, ...)
}
\arguments{
\item{lfun}{A likelihood function, must be definied outside the mymaxlikg function.}

\item{theta}{A sequence between 0 and 1 representing the range in which the probability can realistically lie.}

\item{...}{Extra parameters passed on to plot function.}
}
\value{
Returns a graphical likelihood of a probability based on a supplied likelihood function.
}
\description{
mymaxlikg Creates a graphical likelihood for a probability.
}
\examples{
\dontrun{mymaxlikg(lfun = "logbin2", theta = seq(0,1,length=1000)}
}
