% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mynbin.R
\name{mynbin}
\alias{mynbin}
\title{This function calculates probability of successes based on a negbin distribution.}
\usage{
mynbin(y, r, p)
}
\arguments{
\item{y}{Number of trials}

\item{r}{Number of successes}

\item{p}{Probability of success}
}
\value{
Returns the probability of the number of successes based on a negbin distribution.
}
\description{
This function calculates probability of successes based on a negbin distribution.
}
\examples{
\dontrun{mynbin(10, 3, 0.4))}
}
