% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{This function plots a probability for a normal curve.}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean}

\item{sigma}{Standard Deviation}

\item{a}{Probability upper limit}
}
\value{
Returns a plot and numerical value for the probability of a given number of successes for a normal distribution.
}
\description{
This function plots a probability for a normal curve.
}
\examples{
\dontrun{myncurve(10, 4, 3)}
}
