% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{This generates a scatter plot histogram.}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{An object}

\item{y}{An object}

\item{xlab}{A label}

\item{ylab}{A label}
}
\value{
A scatter plot histogram based on two given parameters and their respective labels.
}
\description{
This generates a scatter plot histogram.
}
\examples{
\dontrun{scatterhist(ddt$LENGTH, ddt$WEIGHT, xlab = "Length", ylab = "Weight")}

}
